Move RoundedBorderDropDownWithLabel component to a dedicated file
