+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/wooshippinglabels/components/RoundedBorderDropdownWithLabel.kt
@@ -1,0 +1,71 @@
+package com.woocommerce.android.ui.orders.wooshippinglabels.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowDropDown
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.woocommerce.android.ui.orders.wooshippinglabels.RoundedCornerBoxWithBorder
+
+
+@Composable
+fun RoundedBorderDropDownWithLabel(
+    label: String,
+    text: String,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Column(modifier = modifier) {
+        Text(
+            text = label,
+            style = MaterialTheme.typography.body2,
+            modifier = Modifier.padding(vertical = 8.dp)
+        )
+        RoundedCornerBoxWithBorder(innerModifier = Modifier.clickable { onClick() }) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(12.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text(
+                    text = text,
+                    style = MaterialTheme.typography.body2,
+                    modifier = Modifier.weight(1f)
+                )
+                Icon(
+                    imageVector = Icons.Filled.ArrowDropDown,
+                    contentDescription = null,
+                    tint = MaterialTheme.colors.onSurface
+                )
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun RoundedBorderDropDownWithLabelPreview() {
+    Column(
+        modifier = Modifier
+            .background(MaterialTheme.colors.background)
+            .padding(16.dp)
+    ) {
+        RoundedBorderDropDownWithLabel(
+            label = "Label",
+            text = "Text",
+            onClick = {}
+        )
+    }
+}
